refactor: extract several maps-api related classes and interfaces to their own files
