+++ src/maps/index.ts
@@ -1,0 +1,4 @@
+export { MapsApiClient } from './MapsApiClient';
+export { MapInstance } from './MapInstance';
+export { MapOptions } from './MapOptions';
+export { MapsDataviews } from './MapsDataviews';
